Consolidates offer service and interface into offer module
