+++ src/app/layout/confirm-box/confirm-box.component.ts
@@ -9,7 +9,6 @@
 import {Component, OnInit, Input} from '@angular/core';
 import { DialogService } from '..//dialog.service';
 import { InfoBoxComponent } from '../info-box/info-box.component';
-import { DataOfferService } from '../../data-management/data-offer.service';
 
 @Component({
   selector: 'rump-confirm-box',
@@ -29,8 +28,7 @@ export class ConfirmBoxComponent implements OnInit {
   @Input() reject: () => void = () => {};
 
 
-  constructor(  private dialogSvc: DialogService,
-                private dataOfferSvc: DataOfferService) { }
+  constructor(  private dialogSvc: DialogService) { }
 
   ngOnInit() {
     this.animateIn = true;
