+++ src/app/layout/side-menu/side-menu.component.ts
@@ -10,7 +10,7 @@ import {Component, OnInit, Output, EventEmitter, Inject} from '@angular/core';
 import { UiStateService, UserService } from '../../services';
 import { DialogBoxComponent } from '../dialog-box/dialog-box.component';
 import { DialogService } from '../dialog.service';
-import { DataOfferService } from '../../data-management/data-offer.service';
+import { DataOfferService } from '../../offers/data-offer.service';
 import { DataPlugService } from '../../data-management/data-plug.service';
 import { MarketSquareService } from '../../market-square/market-square.service';
 import { Subscription, Observable } from 'rxjs/Rx';
