+++ src/app/offers/offer.interface.ts
