--- src/app/offers/offers.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class OffersService {
-
-  constructor() { }
-
-}
